refactor: Renamed component
